pyproject.toml
[project]
name = "chatbot-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "groq>=0.4.0",
    "python-dotenv>=1.0.0",
    "streamlit>=1.30.0",
]
